+++ Cargo.toml
@@ -3,7 +3,7 @@ resolver = "2"
 members = ["api", "program", "utils"]
 
 [workspace.package]
-version = "2.1.5"
+version = "2.1.6"
 edition = "2021"
 license = "Apache-2.0"
 homepage = "https://ore.supply"
