+++ api/src/loaders.rs
@@ -1,4 +1,4 @@
-use ore_utils::{loaders::*, AccountDeserialize, Discriminator};
+use ore_utils::*;
 use solana_program::{account_info::AccountInfo, program_error::ProgramError, pubkey::Pubkey};
 
 use crate::{
