+++ program/src/initialize.rs
@@ -5,7 +5,7 @@ use ore_api::{
     instruction::*,
     state::{Bus, Config, Treasury},
 };
-use ore_utils::{create_pda, loaders::*, spl::create_ata, AccountDeserialize, Discriminator};
+use ore_utils::*;
 use solana_program::{
     account_info::AccountInfo,
     entrypoint::ProgramResult,
