+++ program/src/open.rs
@@ -1,6 +1,7 @@
 use std::mem::size_of;
 
 use ore_api::{consts::*, instruction::OpenArgs, state::Proof};
+use ore_utils::*;
 use solana_program::{
     account_info::AccountInfo,
     clock::Clock,
@@ -11,8 +12,6 @@ use solana_program::{
     system_program,
     sysvar::{self, Sysvar},
 };
-
-use ore_utils::{create_pda, loaders::*, AccountDeserialize, Discriminator};
 
 /// Open creates a new proof account to track a miner's state.
 pub fn process_open(accounts: &[AccountInfo<'_>], data: &[u8]) -> ProgramResult {
