+++ program/src/reset.rs
@@ -4,13 +4,12 @@ use ore_api::{
     loaders::*,
     state::{Bus, Config},
 };
+use ore_utils::*;
 use solana_program::{
     account_info::AccountInfo, clock::Clock, entrypoint::ProgramResult,
     program_error::ProgramError, program_pack::Pack, sysvar::Sysvar,
 };
 use spl_token::state::Mint;
-
-use ore_utils::{loaders::*, AccountDeserialize};
 
 /// Reset tops up the bus balances, updates the base reward rate, and sets up the ORE program for the next epoch.
 pub fn process_reset(accounts: &[AccountInfo<'_>], _data: &[u8]) -> ProgramResult {
