+++ program/src/upgrade.rs
@@ -1,5 +1,5 @@
 use ore_api::{consts::*, error::OreError, instruction::StakeArgs};
-use ore_utils::{loaders::*, spl::mint_to_signed};
+use ore_utils::*;
 use solana_program::{
     account_info::AccountInfo, entrypoint::ProgramResult, program_error::ProgramError,
     program_pack::Pack,
