+++ utils/src/traits.rs
@@ -1,0 +1,10 @@
+use solana_program::program_error::ProgramError;
+
+pub trait Discriminator {
+    fn discriminator() -> u8;
+}
+
+pub trait AccountDeserialize {
+    fn try_from_bytes(data: &[u8]) -> Result<&Self, ProgramError>;
+    fn try_from_bytes_mut(data: &mut [u8]) -> Result<&mut Self, ProgramError>;
+}
